refactor gamma files
